+++ src/main/java/com/aleksander/restaurant/reservation/dto/ReservationSlotDTO.java
@@ -1,0 +1,18 @@
+package com.aleksander.restaurant.reservation.dto;
+
+import java.time.LocalDateTime;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ReservationSlotDTO {
+
+    private LocalDateTime startTime;
+
+    private LocalDateTime endTime;
+
+    private String customerName;
+
+}